Add utility function for checking encoding
- Add basic utility class for functions for use by all data model
  classes
- Currently only a function for checking encoding. This function was
  created to properly encode UTF-8 fields and others with special
characters to UTF-16 for use with nvarchar fields. However, I discovered
this is only necessary with the php7.0-sybase package (or freetds with
php7??) and actually fails with php5-mssql (ie php5-sybase). But I'm
keeping it in case it is needed in the future if we move to php7.